requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
pyyaml
uritemplate
drf-yasg